+++ src/components/Accounts/Actions/ActionModal.js
@@ -446,9 +446,8 @@ class ActionModal extends React.Component {
                     component={Checkbox}
                     type="checkbox"
                     inline
+                    label={<FormattedMessage id="ui-users.accounts.notifyPatron" />}
                   />
-                  {' '}
-                  <FormattedMessage id="ui-users.accounts.notifyPatron" />
                 </Col>
               </Row>
             </div>